+++ packages/contracts/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Allow tokens to be minted to DAO's treasury by passing address(0) as receiver in `TokenFactory`.
+
 ### Changed
 
 - Generalized `MerkleMinter` and made it a `MetaTxComponent`